refactor and delete firebase action

+++ src/routes/Home/index.tsx
@@ -1,23 +1,33 @@
 import { addGoal, removeGoal } from "actions/goalActions";
 import { Icon, Input, List, Spin, Tag } from "antd";
+import { Action } from "redux";
 import * as React from "react";
 import { connect } from "react-redux";
 import { firestoreConnect } from "react-redux-firebase";
 import { bindActionCreators, compose } from "redux";
 import { ThunkDispatch } from "redux-thunk";
 import { AppState } from "store";
-import { GoalActionTypes } from "types/actions";
 import { IGoal } from "types/goal";
 import uuidv4 from "uuid/v4";
 import styles from "./home.module.scss";
+import { IGoalState } from "types/goal";
 
 const { Search } = Input;
 
 interface HomePageProps {
-  goalsData: IGoal[];
+  goalsData: IGoalState;
 }
 
 interface HomePageState {}
+
+interface LinkStateProps {
+  goalsData: IGoalState;
+}
+
+interface LinkDispatchProps {
+  addGoal: (goal: IGoal) => void;
+  removeGoal: (id: string) => void;
+}
 
 type Props = HomePageProps & LinkDispatchProps;
 type State = HomePageState & LinkStateProps;
@@ -25,6 +35,8 @@ type State = HomePageState & LinkStateProps;
 class Home extends React.Component<Props, State> {
   public render() {
     const { goalsData } = this.props;
+
+    const goalsIds = goalsData && Object.keys(goalsData);
 
     return (
       <div className={styles.homeWrapper}>
@@ -40,22 +52,31 @@ class Home extends React.Component<Props, State> {
         </div>
         <div className={styles.goalsContainer}>
           {goalsData ? (
-            goalsData.map(goal => (
-              <div className={styles.goalCard} key={goal.id}>
-                <div className={styles.goalHeader}>
-                  <span className={styles.goalName}>{goal.name}</span>
-                  <Tag key={goal.id + 1} color={goal.paid ? "green" : "red"}>
-                    {goal.paid ? "paid" : "not paid"}
-                  </Tag>
-                  <Icon
-                    type="delete"
-                    onClick={() => this.deleteGoal(goal.id)}
-                    className={styles.deleteIcon}
-                  />
-                </div>
-                <Icon type="down-circle" className={styles.more} />
-              </div>
-            ))
+            goalsIds.map(goalId => {
+              const goalData = goalsData[goalId];
+              if (goalData) {
+                return (
+                  <div className={styles.goalCard} key={goalId}>
+                    <div className={styles.goalHeader}>
+                      <span className={styles.goalName}>{goalData.name}</span>
+                      <Tag
+                        key={goalData.id + 1}
+                        color={goalData.paid ? "green" : "red"}
+                      >
+                        {goalData.paid ? "paid" : "not paid"}
+                      </Tag>
+                      <Icon
+                        type="delete"
+                        onClick={() => this.deleteGoal(goalId)}
+                        className={styles.deleteIcon}
+                      />
+                    </div>
+                    <Icon type="down-circle" className={styles.more} />
+                  </div>
+                );
+              }
+              return null;
+            })
           ) : (
             <Spin size="large" />
           )}
@@ -82,23 +103,12 @@ class Home extends React.Component<Props, State> {
   };
 }
 
-interface LinkStateProps {
-  goalsData: IGoal[];
-}
-
-interface LinkDispatchProps {
-  addGoal: (goal: IGoal) => void;
-  removeGoal: (id: string) => void;
-}
-
 export default compose<any>(
   connect(
     (state: AppState): LinkStateProps => ({
-      goalsData: state.firestore.ordered.Goals
+      goalsData: state.firestore.data.Goals
     }),
-    (
-      dispatch: ThunkDispatch<any, any, GoalActionTypes>
-    ): LinkDispatchProps => ({
+    (dispatch: ThunkDispatch<any, any, Action>): LinkDispatchProps => ({
       addGoal: bindActionCreators(addGoal, dispatch),
       removeGoal: bindActionCreators(removeGoal, dispatch)
     })
